More separation of server from engine
Client starts
Working on protocol
