+++ src/main/java/org/spout/engine/world/SpoutServerWorld.java
@@ -32,6 +32,7 @@ import java.lang.ref.WeakReference;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
+import java.util.Set;
 import java.util.UUID;
 
 import org.spout.api.Server;
@@ -49,15 +50,19 @@ import org.spout.api.geo.discrete.Transform;
 import org.spout.api.io.bytearrayarray.BAAWrapper;
 import org.spout.api.math.Quaternion;
 import org.spout.api.math.Vector3;
+import org.spout.api.scheduler.TaskManager;
 import org.spout.api.util.StringMap;
 import org.spout.api.util.list.concurrent.ConcurrentList;
 import org.spout.api.util.list.concurrent.setqueue.SetQueue;
 import org.spout.api.util.map.WeakValueHashMap;
 import org.spout.engine.SpoutEngine;
 import org.spout.engine.filesystem.versioned.WorldFiles;
+import org.spout.engine.scheduler.SpoutScheduler;
+import org.spout.engine.scheduler.SpoutTaskManager;
+import org.spout.engine.util.thread.AsyncManager;
 import org.spout.engine.util.thread.snapshotable.SnapshotableLong;
 
-public class SpoutServerWorld extends SpoutWorld {
+public class SpoutServerWorld extends SpoutWorld implements AsyncManager {
 	/**
 	 * The spawn position.
 	 */
@@ -86,6 +91,11 @@ public class SpoutServerWorld extends SpoutWorld {
 	 * RegionFile manager for the world
 	 */
 	private final RegionFileManager regionFileManager;
+	protected final SpoutTaskManager taskManager;
+	/**
+	 * The execution thread for this world
+	 */
+	private Thread executionThread;
 	/*
 	 * A WeakReference to this world
 	 */
@@ -109,6 +119,8 @@ public class SpoutServerWorld extends SpoutWorld {
 		this.age = new SnapshotableLong(snapshotManager, age);
 		spawnLocation.set(new Transform(new Point(this, 1, 100, 1), Quaternion.IDENTITY, Vector3.ONE));
 		selfReference = new WeakReference<SpoutServerWorld>(this);
+		
+		taskManager = new SpoutTaskManager(getEngine().getScheduler(), null, this, age);
 
 		getEngine().getScheduler().addAsyncManager(this);
 	}
@@ -125,20 +137,6 @@ public class SpoutServerWorld extends SpoutWorld {
 		}
 		SpoutServerWorld world = (SpoutServerWorld) obj;
 		return world.getUID().equals(getUID());
-	}
-
-	@Override
-	public void startTickRun(int stage, long delta) {
-		switch (stage) {
-			case 0: {
-				age.set(age.get() + delta);
-				super.startTickRun(stage, delta);
-				break;
-			}
-			default: {
-				throw new IllegalStateException("Number of states exceeded limit for SpoutWorld");
-			}
-		}
 	}
 
 	@Override
@@ -269,5 +267,99 @@ public class SpoutServerWorld extends SpoutWorld {
 
 	public StringMap getLightingMap() {
 		return lightingMap;
+	}
+	
+	@Override
+	public TaskManager getTaskManager() {
+		return taskManager;
+	}
+
+	@Override
+	public void copySnapshotRun() {
+		synchronized (regionColumnDirtyQueueMap) {
+			// This performs copy snapshot and also clears the column dirty queues
+			Set<Long> keys = regionColumnDirtyQueueMap.keySet();
+			for (Long key : keys) {
+				SetQueue<SpoutColumn> queue = regionColumnDirtyQueueMap.safeGet(key);
+				if (queue != null) {
+					SpoutColumn col;
+					while ((col = queue.poll()) != null) {
+						col.copySnapshot();
+					}
+				}
+			}
+			regionColumnDirtyQueueMap.flushKeys();
+		}
+		snapshotManager.copyAllSnapshots();
+	}
+
+	@Override
+	public void startTickRun(int stage, long delta) {
+		switch (stage) {
+			case 0: {
+				age.set(age.get() + delta);
+				parallelTaskManager.heartbeat(delta);
+				taskManager.heartbeat(delta);
+				for (Component component : values()) {
+					component.tick(delta);
+				}
+				break;
+			}
+			default: {
+				throw new IllegalStateException("Number of states exceeded limit for SpoutWorld");
+			}
+		}
+	}
+
+	@Override
+	public int getMaxStage() {
+		return 0;
+	}
+
+	@Override
+	public void finalizeRun() {
+		synchronized (columnSet) {
+			for (SpoutColumn c : columnSet) {
+				c.onFinalize();
+			}
+		}
+	}
+
+	@Override
+	public void preSnapshotRun() {
+
+	}
+
+	// Worlds don't do any of these
+	@Override
+	public void runPhysics(int sequence) {
+	}
+
+	@Override
+	public void runLighting(int sequence) {
+	}
+
+	@Override
+	public long getFirstDynamicUpdateTime() {
+		return SpoutScheduler.END_OF_THE_WORLD;
+	}
+
+	@Override
+	public void runDynamicUpdates(long time, int sequence) {
+	}
+
+	@Override
+	public int getSequence() {
+		return 0;
+	}
+
+	@Override
+	public Thread getExecutionThread() {
+		return executionThread;
+	}
+
+	@Override
+	public void setExecutionThread(Thread t) {
+		this.executionThread = t;
 	}
 }
